Refactoring metrics in RestServer
